Refactored EnsurePackageOperation to remove most of the #conditionals
