+++ InedoCore/OtterExtension/Operations/ProGet/CollectPackagesOperation.cs
@@ -1,9 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
+using Inedo.Diagnostics;
 using Inedo.Documentation;
 using Inedo.Extensions.UniversalPackages;
 using Inedo.Otter.Data;
@@ -21,27 +19,49 @@ namespace Inedo.Extensions.Operations.ProGet
     {
         public override async Task<PersistedConfiguration> CollectAsync(IOperationExecutionContext context)
         {
+            IList<RegisteredPackage> packages;
+            this.LogDebug("Connecting to machine package registry...");
             using (var registry = await PackageRegistry.GetRegistryAsync(context.Agent, false).ConfigureAwait(false))
             {
+                this.LogDebug("Acquiring package registry lock...");
                 await registry.LockAsync(context.CancellationToken).ConfigureAwait(false);
+                this.LogDebug($"Package registry lock acquired (token={registry.LockToken}).");
 
-                foreach (var p in await registry.GetInstalledPackagesAsync().ConfigureAwait(false))
-                {
-#warning this does not handle uninstalled packages
-                    DB.ServerPackages_CreateOrUpdatePackage(
-                        Server_Id: context.ServerId,
-                        PackageType_Name: "ProGet",
-                        Package_Name: p.Name,
-                        Package_Version: p.Version,
-                        CollectedOn_Execution_Id: context.ExecutionId,
-                        Url_Text: p.FeedUrl
-                    );
-                }
+                this.LogInformation("Retreiving list of packages...");
+                packages = await registry.GetInstalledPackagesAsync().ConfigureAwait(false);
+                this.LogInformation("Packages installed: " + packages.Count);
 
                 // doesn't need to be in a finally because dispose will unlock if necessary, but prefer doing it asynchronously
                 await registry.UnlockAsync().ConfigureAwait(false);
             }
 
+            this.LogDebug("Recording installed packages...");
+            using (var db = new DB.Context())
+            {
+                db.BeginTransaction();
+
+                await db.ServerPackages_DeletePackagesAsync(
+                    Server_Id: context.ServerId,
+                    PackageType_Name: "ProGet"
+                ).ConfigureAwait(false);
+
+                foreach (var p in packages)
+                {
+                    await db.ServerPackages_CreateOrUpdatePackageAsync(
+                        Server_Id: context.ServerId,
+                        PackageType_Name: "ProGet",
+                        Package_Name: p.Name,
+                        Package_Version: p.Version,
+                        CollectedOn_Execution_Id: context.ExecutionId,
+                        Url_Text: p.FeedUrl,
+                        CollectedFor_ServerRole_Id: context.ServerRoleId
+                    ).ConfigureAwait(false);
+                }
+
+                db.CommitTransaction();
+            }
+
+            this.LogInformation("Package collection complete.");
             return null;
         }
